requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy>=1.26.0
pydantic>=2.0.0
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=0.19.0
